+++ R/errors.R
@@ -158,3 +158,8 @@ renv_error_tag <- function(e) {
 renv_error_handler_call <- function() {
   as.call(list(renv_error_handler))
 }
+
+renv_error_report <- function(error = NULL) {
+  if (renv_tests_running() && inherits(error, "error"))
+    stop(error)
+}