+++ tests/testthat/test-dependencies.R
@@ -22,16 +22,12 @@ test_that("usages of library, etc. are properly handled", {
 })
 
 test_that("parse errors are okay in .Rmd documents", {
-  skip_if_not_installed("knitr")
-  skip_if_not_installed("rmarkdown")
   deps <- dependencies("resources/chunk-errors.Rmd", quiet = TRUE)
   pkgs <- deps$Package
   expect_setequal(pkgs, c("rmarkdown", "dplyr"))
 })
 
 test_that("inline chunks are parsed for dependencies", {
-  skip_if_not_installed("knitr")
-  skip_if_not_installed("rmarkdown")
   deps <- dependencies("resources/inline-chunks.Rmd")
   pkgs <- deps$Package
   expect_setequal(pkgs, c("rmarkdown", "inline", "multiple", "separate"))
@@ -51,8 +47,6 @@ test_that("the package name is validated when inferring dependencies", {
 })
 
 test_that("empty chunks don't cause issues during dependency resolution", {
-  skip_if_not_installed("knitr")
-  skip_if_not_installed("rmarkdown")
   deps <- dependencies("resources/empty-chunk.Rmd")
   pkgs <- deps$Package
   expect_setequal(pkgs, c("rmarkdown"))
@@ -103,8 +97,6 @@ test_that("renv warns when large number of files found", {
 })
 
 test_that("evil knitr chunks are handled", {
-  skip_if_not_installed("knitr")
-  skip_if_not_installed("rmarkdown")
   deps <- dependencies("resources/evil.Rmd")
   packages <- deps$Package
   expect_setequal(packages, c("rmarkdown", "a", "b"))
@@ -147,9 +139,6 @@ test_that("dependencies can infer the root directory", {
 
 test_that("no warnings are produced when crawling dependencies", {
 
-  skip_if_not_installed("knitr")
-  skip_if_not_installed("rmarkdown")
-
   expect_warning(
     regexp = NA,
     dependencies(
@@ -185,9 +174,6 @@ test_that("packages referenced by modules::import() are discovered", {
 })
 
 test_that("dependencies specified in R Markdown site generators are found", {
-
-  skip_if_not_installed("knitr")
-  skip_if_not_installed("rmarkdown")
 
   renv_tests_scope()
   writeLines(
@@ -354,6 +340,12 @@ test_that("we parse package references from arbitrary yaml fields", {
   deps <- dependencies("resources/rmd-base-format.Rmd", progress = FALSE)
   expect_true("bookdown" %in% deps$Package)
   expect_true("rticles" %in% deps$Package)
+})
+
+test_that("dependencies in parameterized documents are discovered", {
+  deps <- dependencies("resources/params.Rmd", progress = FALSE)
+  expect_true(all(c("shiny", "A") %in% deps$Package))
+  expect_false("B" %in% deps$Package)
 })
 
 test_that("dependencies in hidden folders are not scoured", {